15-05-2026_2

+++ crud/shoppiq/src/main/java/com/crud/project/shoppiq/services/ItemService.java
@@ -20,17 +20,17 @@ public class ItemService {
     }
 
     @Transactional
-    public Optional<Item> getItemById(int id) {
+    public Optional<Item> getItemById(long id) {
         return itemRepository.findById(id);
     }
 
     @Transactional
-    public void deleteItemById(int id) {
+    public void deleteItemById(long id) {
         itemRepository.deleteById(id);
     }
 
     @Transactional
-    public Optional<Item> updateItemById(int id, Item newItem) {
+    public Optional<Item> updateItemById(long id, Item newItem) {
         return itemRepository.updateById(id, newItem);
     }
 }
